changed default behaviour for the configure wrapper outside of the build dir
Calling configure in a directory which is not manged by builddir now executes
the configure in the current working directory instead of erroring out.

The old behaviour broke the native extension build of libxml2 in the nokogiri gem.
The build script executed configure (without ./) in the build directory which
failed. The new behaviour leads to the expected result by forwarding the call
to the configure script.